# test/openapi.json
{"openapi":"3.0.0","paths":{"/api":{"get":{"operationId":"AppController_getHello","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"string"}}}}},"tags":["App"]}},"/api/auth/login":{"post":{"operationId":"AuthController_login","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginDto"}}}},"responses":{"201":{"description":""}},"tags":["Auth"]}},"/api/auth/me":{"get":{"operationId":"AuthController_me","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"tags":["Auth"]}},"/api/auth/logout":{"post":{"operationId":"AuthController_logout","parameters":[],"responses":{"201":{"description":""}},"tags":["Auth"]}},"/api/auth/google":{"get":{"operationId":"AuthController_google","parameters":[],"responses":{"200":{"description":""}},"tags":["Auth"]}},"/api/auth/google/callback":{"get":{"operationId":"AuthController_googleCallback","parameters":[],"responses":{"200":{"description":""}},"tags":["Auth"]}},"/api/users/assistants":{"get":{"operationId":"UsersController_getAssistants","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"type":"object"}}}}}},"tags":["Users"]}},"/api/users/editors":{"get":{"operationId":"UsersController_getEditors","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"type":"object"}}}}}},"tags":["Users"]}},"/api/dashboard/summary":{"get":{"operationId":"DashboardController_summary","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"tags":["Dashboard"]}},"/api/dashboard/series":{"get":{"operationId":"DashboardController_series","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"type":"object"}}}}}},"tags":["Dashboard"]}},"/api/dashboard/tasks":{"get":{"operationId":"DashboardController_tasks","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"type":"object"}}}}}},"tags":["Dashboard"]}},"/api/dashboard/submissions":{"get":{"operationId":"DashboardController_submissions","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"type":"object"}}}}}},"tags":["Dashboard"]}},"/api/dashboard/notifications":{"get":{"operationId":"DashboardController_notifications","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"type":"object"}}}}}},"tags":["Dashboard"]}},"/api/uploads":{"post":{"operationId":"UploadsController_uploadFile","parameters":[],"responses":{"201":{"description":""}},"tags":["Uploads"]}},"/api/notifications":{"get":{"operationId":"NotificationsController_list","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"type":"object"}}}}}},"tags":["Notifications"]}},"/api/notifications/read-all":{"patch":{"operationId":"NotificationsController_readAll","parameters":[],"responses":{"200":{"description":""}},"tags":["Notifications"]}},"/api/notifications/{id}/read":{"patch":{"operationId":"NotificationsController_read","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Notifications"]}},"/api/proposals":{"post":{"operationId":"ProposalsController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateProposalDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"tags":["Proposals"]}},"/api/proposals/mine":{"get":{"operationId":"ProposalsController_listMine","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"type":"object"}}}}}},"tags":["Proposals"]}},"/api/proposals/{id}/submit":{"patch":{"operationId":"ProposalsController_submit","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"tags":["Proposals"]}},"/api/proposals/review-queue":{"get":{"operationId":"ProposalsController_reviewQueue","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"type":"object"}}}}}},"tags":["Proposals"]}},"/api/proposals/{id}/decision":{"patch":{"operationId":"ProposalsController_decide","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DecisionDto"}}}},"responses":{"200":{"description":""}},"tags":["Proposals"]}},"/api/genres":{"get":{"operationId":"GenresController_getAll","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"type":"object"}}}}}},"tags":["Genres"]}},"/api/series/all":{"get":{"operationId":"SeriesController_listAll","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"type":"object"}}}}}},"tags":["Series"]}},"/api/series/mine":{"get":{"operationId":"SeriesController_listMine","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"type":"object"}}}}}},"tags":["Series"]}},"/api/series/{id}":{"get":{"operationId":"SeriesController_getOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"tags":["Series"]}},"/api/series/{id}/editor":{"put":{"operationId":"SeriesController_assignEditor","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AssignEditorDto"}}}},"responses":{"200":{"description":""}},"tags":["Series"]},"delete":{"operationId":"SeriesController_unassignEditor","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Series"]}},"/api/chapters":{"post":{"operationId":"ChaptersController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateChapterDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"tags":["Chapters"]},"get":{"operationId":"ChaptersController_listBySeries","parameters":[{"name":"seriesId","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"type":"object"}}}}}},"tags":["Chapters"]}},"/api/chapters/{id}/status":{"patch":{"operationId":"ChaptersController_setStatus","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChapterStatusDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"tags":["Chapters"]}},"/api/chapters/review-queue":{"get":{"operationId":"ChaptersController_reviewQueue","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"type":"object"}}}}}},"tags":["Chapters"]}},"/api/chapters/{id}/pages":{"get":{"operationId":"ChaptersController_editorPages","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"type":"object"}}}}}},"tags":["Chapters"]}},"/api/chapters/{id}/editor-review":{"patch":{"operationId":"ChaptersController_editorReview","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/EditorReviewDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"tags":["Chapters"]}},"/api/pages":{"post":{"operationId":"PagesController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreatePageDto"}}}},"responses":{"201":{"description":""}},"tags":["Pages"]},"get":{"operationId":"PagesController_listByChapter","parameters":[{"name":"chapterId","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"type":"object"}}}}}},"tags":["Pages"]}},"/api/pages/{id}":{"get":{"operationId":"PagesController_getOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Pages"]}},"/api/regions":{"post":{"operationId":"RegionsController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateRegionDto"}}}},"responses":{"201":{"description":""}},"tags":["Regions"]},"get":{"operationId":"RegionsController_listByPage","parameters":[{"name":"pageId","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"type":"object"}}}}}},"tags":["Regions"]}},"/api/regions/{id}":{"delete":{"operationId":"RegionsController_delete","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Regions"]}},"/api/tasks":{"post":{"operationId":"TasksController_assign","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateTaskDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"tags":["Tasks"]},"get":{"operationId":"TasksController_listByPage","parameters":[{"name":"pageId","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"type":"object"}}}}}},"tags":["Tasks"]}},"/api/tasks/mine":{"get":{"operationId":"TasksController_listMine","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"type":"object"}}}}}},"tags":["Tasks"]}},"/api/tasks/{id}/start":{"patch":{"operationId":"TasksController_start","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"tags":["Tasks"]}},"/api/submissions":{"post":{"operationId":"SubmissionsController_submit","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateSubmissionDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"tags":["Submissions"]}},"/api/submissions/review-queue":{"get":{"operationId":"SubmissionsController_reviewQueue","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"type":"object"}}}}}},"tags":["Submissions"]}},"/api/submissions/{id}/review":{"patch":{"operationId":"SubmissionsController_review","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReviewSubmissionDto"}}}},"responses":{"200":{"description":""}},"tags":["Submissions"]}},"/api/studio/page-versions":{"post":{"operationId":"StudioController_savePageVersion","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SavePageVersionDto"}}}},"responses":{"201":{"description":""}},"tags":["Studio"]}},"/api/studio/docs":{"post":{"operationId":"StudioController_saveDoc","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SaveDocDto"}}}},"responses":{"201":{"description":""}},"tags":["Studio"]}},"/api/studio/docs/{pageId}":{"get":{"operationId":"StudioController_getDoc","parameters":[{"name":"pageId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"tags":["Studio"]}},"/api/admin/users":{"get":{"operationId":"AdminController_users","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"type":"object"}}}}}},"tags":["Admin"]}},"/api/admin/overview":{"get":{"operationId":"AdminController_overview","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"tags":["Admin"]}},"/api/admin/users/{id}":{"patch":{"operationId":"AdminController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUserDto"}}}},"responses":{"200":{"description":""}},"tags":["Admin"]}},"/api/vote-periods":{"post":{"operationId":"RankingsController_openPeriod","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateVotePeriodDto"}}}},"responses":{"201":{"description":""}},"tags":["Rankings"]}},"/api/vote-periods/open":{"get":{"operationId":"RankingsController_openPeriods","parameters":[],"responses":{"200":{"description":""}},"tags":["Rankings"]}},"/api/votes":{"post":{"operationId":"RankingsController_castVote","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateVoteDto"}}}},"responses":{"201":{"description":""}},"tags":["Rankings"]}},"/api/vote-periods/{id}/close":{"post":{"operationId":"RankingsController_closePeriod","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"201":{"description":""}},"tags":["Rankings"]}},"/api/rankings":{"get":{"operationId":"RankingsController_leaderboard","parameters":[],"responses":{"200":{"description":""}},"tags":["Rankings"]}},"/api/decisions":{"post":{"operationId":"DecisionsController_decide","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateDecisionDto"}}}},"responses":{"201":{"description":""}},"tags":["Decisions"]},"get":{"operationId":"DecisionsController_listForSeries","parameters":[{"name":"seriesId","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"type":"object"}}}}}},"tags":["Decisions"]}},"/api/annotations":{"post":{"operationId":"AnnotationsController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateAnnotationDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"tags":["Annotations"]},"get":{"operationId":"AnnotationsController_list","parameters":[{"name":"targetType","required":true,"in":"query","schema":{"type":"string"}},{"name":"targetId","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"type":"object"}}}}}},"tags":["Annotations"]}},"/api/annotations/{id}/resolve":{"patch":{"operationId":"AnnotationsController_resolve","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Annotations"]}},"/api/earnings/mine":{"get":{"operationId":"EarningsController_mine","parameters":[],"responses":{"200":{"description":""}},"tags":["Earnings"]}},"/api/disputes":{"post":{"operationId":"DisputesController_file","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateDisputeDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"tags":["Disputes"]},"get":{"operationId":"DisputesController_listAll","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"type":"object"}}}}}},"tags":["Disputes"]}},"/api/disputes/mine":{"get":{"operationId":"DisputesController_mine","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"type":"object"}}}}}},"tags":["Disputes"]}},"/api/disputes/{id}/review":{"patch":{"operationId":"DisputesController_markUnderReview","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Disputes"]}},"/api/disputes/{id}/resolve":{"patch":{"operationId":"DisputesController_resolve","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResolveDisputeDto"}}}},"responses":{"200":{"description":""}},"tags":["Disputes"]}}},"info":{"title":"Manga Creation Workflow & Publishing Management System API","description":"Internal manga-studio production & publishing platform — REST API. Authenticate via POST /api/auth/login, then click Authorize and paste the Bearer token.","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"LoginDto":{"type":"object","properties":{"email":{"type":"string","format":"email"},"password":{"type":"string","minLength":6}},"required":["email","password"]},"CreateProposalDto":{"type":"object","properties":{"title":{"type":"string"},"synopsis":{"type":"string"},"proposedFrequency":{"type":"string","enum":["WEEKLY","MONTHLY"]},"genreIds":{"minItems":1,"type":"array","items":{"type":"number"}}},"required":["title","proposedFrequency","genreIds"]},"DecisionDto":{"type":"object","properties":{"decision":{"type":"string","enum":["APPROVED","REJECTED"]}},"required":["decision"]},"AssignEditorDto":{"type":"object","properties":{"editorUserId":{"type":"number","minimum":1}},"required":["editorUserId"]},"CreateChapterDto":{"type":"object","properties":{"seriesId":{"type":"number"},"title":{"type":"string"},"deadline":{"type":"string"}},"required":["seriesId","title"]},"ChapterStatusDto":{"type":"object","properties":{"status":{"type":"string","enum":["DRAFT","IN_PROGRESS","READY_FOR_EDITOR_REVIEW","EDITOR_APPROVED","PUBLISHED"]}},"required":["status"]},"EditorReviewDto":{"type":"object","properties":{"decision":{"type":"string","enum":["APPROVE","REVISE"]},"feedback":{"type":"string"}},"required":["decision"]},"CreatePageDto":{"type":"object","properties":{"chapterId":{"type":"number"},"imageUrl":{"type":"string"},"uploadNote":{"type":"string"}},"required":["chapterId","imageUrl"]},"CreateRegionDto":{"type":"object","properties":{"pageId":{"type":"number"},"regionType":{"type":"string","enum":["PANEL","BACKGROUND","CHARACTER","DIALOGUE_BUBBLE","EFFECT"]},"x":{"type":"number"},"y":{"type":"number"},"width":{"type":"number"},"height":{"type":"number"}},"required":["pageId","regionType","x","y","width","height"]},"CreateTaskDto":{"type":"object","properties":{"regionId":{"type":"number"},"assigneeUserId":{"type":"number"},"description":{"type":"string"},"instruction":{"type":"string"},"deadline":{"type":"string"}},"required":["regionId","assigneeUserId"]},"CreateSubmissionDto":{"type":"object","properties":{"taskId":{"type":"number"},"fileUrl":{"type":"string"},"versionNote":{"type":"string"}},"required":["taskId","fileUrl"]},"ReviewSubmissionDto":{"type":"object","properties":{"decision":{"type":"string","enum":["APPROVED","REVISION_REQUIRED"]},"feedback":{"type":"string"}},"required":["decision"]},"SavePageVersionDto":{"type":"object","properties":{"pageId":{"type":"number"},"imageUrl":{"type":"string"},"note":{"type":"string"}},"required":["pageId","imageUrl"]},"SaveDocDto":{"type":"object","properties":{"pageId":{"type":"number"},"manifest":{"type":"object"}},"required":["pageId","manifest"]},"UpdateUserDto":{"type":"object","properties":{"isActivated":{"type":"boolean"},"role":{"type":"string","enum":["MANGAKA","ASSISTANT","TANTOU_EDITOR","EDITORIAL_BOARD","ADMIN"]}}},"CreateVotePeriodDto":{"type":"object","properties":{"seriesId":{"type":"number"},"periodType":{"type":"string","enum":["WEEKLY","MONTHLY"]},"startDate":{"type":"string"},"endDate":{"type":"string"}},"required":["seriesId","periodType","startDate","endDate"]},"CreateVoteDto":{"type":"object","properties":{"votePeriodId":{"type":"number"},"score":{"type":"number","minimum":1,"maximum":5},"comment":{"type":"string"}},"required":["votePeriodId","score"]},"CreateDecisionDto":{"type":"object","properties":{"seriesId":{"type":"number"},"decisionType":{"type":"string","enum":["CONTINUE","CANCEL","CHANGE_FREQUENCY","HIATUS"]},"newFrequency":{"type":"string","enum":["WEEKLY","MONTHLY"]},"reason":{"type":"string"}},"required":["seriesId","decisionType","reason"]},"CreateAnnotationDto":{"type":"object","properties":{"targetType":{"type":"string","enum":["PAGE","MANUSCRIPT","SUBMISSION"]},"targetId":{"type":"number"},"category":{"type":"string","enum":["CONTENT_ISSUE","DIALOGUE_ISSUE","SCRIPT_ISSUE","VISUAL_ISSUE","GENERAL"]},"context":{"type":"string"},"x":{"type":"number"},"y":{"type":"number"}},"required":["targetType","targetId","category","context"]},"CreateDisputeDto":{"type":"object","properties":{"taskId":{"type":"number"},"reason":{"type":"string"},"expectedAmount":{"type":"number"}},"required":["taskId","reason"]},"ResolveDisputeDto":{"type":"object","properties":{"status":{"type":"string","enum":["RESOLVED","REJECTED"]},"resolutionNote":{"type":"string"},"adjustedAmount":{"type":"number"}},"required":["status","resolutionNote"]}}}}